% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_taxon_attribute.R
\name{is_taxonlevel}
\alias{is_taxonlevel}
\title{Test of een taxon een bepaald taxonlevel heeft}
\usage{
is_taxonlevel(taxa, taxonlevel = NULL)
}
\arguments{
\item{taxa}{Een vector met taxonnamen}

\item{taxonlevel}{Het taxonlevel dat getest moet worden. Als het niet wordt opgegeven dan
wordt het taxonlevel op "Species gezet". Zie \code{twn::taxonlevels} voor alle mogelijke taxonlevels.
Het taxonlevel is niet hoofdlettergevoelig}
}
\value{
Een logische vector. Als een taxon ontbreekt of niet in de TWN-lijst voorkomt wordt \code{NA}
geretourneerd.
}
\description{
Deze functie kijkt of een taxon een bepaald taxonlevel heeft en retourneert een logische vector.
}
\examples{
is_taxonlevel(c("Bufo bufo", "Bufo", "Bufonidae", "Buf", NA), "Species")

}
