% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_taxon_attribute.R
\name{is_taxontype}
\alias{is_taxontype}
\title{Test of een taxon een bepaald taxontype heeft}
\usage{
is_taxontype(taxa, taxontype)
}
\arguments{
\item{taxa}{Een vector met taxonnamen}

\item{taxontype}{Het taxontype dat getest moet worden. Zie \code{unique(twn_lijst$taxontype)}
voor alle mogelijke taxontypes. Het taxontype is niet hoofdlettergevoelig}
}
\value{
Een logische vector. Als een taxon ontbreekt of niet in de TWN-lijst voorkomt wordt \code{NA}
geretourneerd.
}
\description{
Deze functie kijkt of een taxon een bepaald taxontype heeft en retourneert een logische vector.
}
\examples{
is_taxontype(c("Bufo bufo", "Abies", "Bufonidae", "Buf", NA), "Amphibia")

}
