% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/krw_mafy_doc.R
\docType{data}
\name{krw_mafy_categorie}
\alias{krw_mafy_categorie}
\title{KRW-maatlatten: Macrofyten categorieen}
\format{Dataframe
\itemize{
\item \code{groep} Een aanduiding of het taxon behoort tot de \emph{Helofyten} of \emph{Hydrofyten}.
Dit onderscheid wordt alleen gemaakt voor de watertypen M1 t/m M10 in de
maatlatten van 2018
\item \code{biotaxon.naam} Naam van het biotaxon, inclusief ondersoorten die niet expliciet in het
maatlatdocument zijn opgenomen.
\item \code{biotaxon.naam_bron} Naam van het biotaxon volgens het maatlatdocument.
\item \code{krwwatertype.code} KRW-watertypecode
\item \code{categorie} Categorie van het taxon voor het betreffende watertype. Waarde 1 t/m 5.
}}
\source{
\href{https://www.aquo-kit.nl/}{Aquo-kit}
}
\usage{
krw_mafy_categorie
}
\description{
Een van drie tabellen die benodigd zijn voor de KRW-toetsing van macrofyten.
}
\details{
De macrofyten kunnen getoetst worden voor de Kaderrichtlijn Water d.m.v. zogenaamde maatlatten.
De documentatie voor deze maatlatten kan gevonden worden op de website van \href{www.stowa.nl}{STOWA}.
Het betreft:
\itemize{
\item Referenties en maatlatten voor natuurlijke watertypen voor de Kaderrichtlijn Water 2015-2021
\item Omschrijving MEP en maatlatten voor sloten en kanalen voor de Kaderrichtlijn Water 2015-2021
\item Referenties en maatlatten voor natuurlijke watertypen voor de Kaderrichtlijn Water 2021-2027
\item Omschrijving MEP en maatlatten voor sloten en kanalen voor de Kaderrichtlijn Water 2021-2027
}
}
\note{
In deze tabel zijn zowel de categorieën van de maatlat van 2012 als van 2018 opgenomen.
}
\section{Metadata dataset}{

\itemize{
\item \strong{Naam:} Macrofyten-categorieen
\item \strong{Oorspronkelijke Naam:} 250aMacrofyten-categorieen_yyyymmdduummss
\item \strong{Bronhouder:} IHW (voor Aquo-kit) en STOWA voor de onderliggende data
\item \strong{Herkomst:} Aquo-kit \url{https://www.aquo-kit.nl/}
\item \strong{Auteurs:} Diverse
\item \strong{Datum dataset:} 2018
\item \strong{Datum opname dataset:} 25-6-2019
\item \strong{Gebruiksrechten:} Onbeperkt
\item \strong{Soortgroep:} Macrofyten
}
}

\section{Bewerkingen}{


De kolomnamen zijn omgezet naar lowercase.
}

\examples{

head(krw_mafy_categorie)
}
\seealso{
Zie ook \link{krw_mafy_score} en \link{krw_mafy_constanten}.
}
\keyword{datasets}
