% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/krw_mafy_doc.R
\docType{data}
\name{krw_mafy_score}
\alias{krw_mafy_score}
\title{KRW-maatlatten: Macrofyten scores}
\format{Dataframe
\itemize{
\item \code{groep} Een aanduiding of het taxon behoort tot de \emph{Helofyten} of \emph{Hydrofyten}.
Dit onderscheid wordt alleen gemaakt voor de watertypen M1 t/m M10 in de
maatlatten van 2018
\item \code{krwwatertype.code} KRW-watertypecode
\item \code{abundantieklasse} Klasse die de vegetatieabundantie weergeeft. Waarde 1 t/m 3.
\item \code{categorie} Categorie van het taxon voor het betreffende watertype. Waarde 1 t/m 5.
\item \code{score} Score die aan een taxon wordt toegekend o.b.v. de categorie en de abundantie
van het taxon.
}}
\source{
\href{https://www.aquo-kit.nl/}{Aquo-kit}
}
\usage{
krw_mafy_score
}
\description{
Een van drie tabellen die benodigd zijn voor de KRW-toetsing van macrofyten.
}
\details{
De macrofyten kunnen getoetst worden voor de Kaderrichtlijn Water d.m.v. zogenaamde maatlatten.
De documentatie voor deze maatlatten kan gevonden worden op de website van \href{www.stowa.nl}{STOWA}.
Het betreft:
\itemize{
\item Referenties en maatlatten voor natuurlijke watertypen voor de Kaderrichtlijn Water 2015-2021
\item Omschrijving MEP en maatlatten voor sloten en kanalen voor de Kaderrichtlijn Water 2015-2021
\item Referenties en maatlatten voor natuurlijke watertypen voor de Kaderrichtlijn Water 2021-2027
\item Omschrijving MEP en maatlatten voor sloten en kanalen voor de Kaderrichtlijn Water 2021-2027
}
}
\section{Metadata dataset}{

\itemize{
\item \strong{Naam:} Macrofyten-scores
\item \strong{Oorspronkelijke Naam:} 250cMacrofyten-scores_yyyymmdduummss
\item \strong{Bronhouder:} IHW (voor Aquo-kit) en STOWA voor de onderliggende data
\item \strong{Herkomst:} Aquo-kit \url{https://www.aquo-kit.nl/}
\item \strong{Auteurs:} Diverse
\item \strong{Datum dataset:} 2018
\item \strong{Datum opname dataset:} 25-6-2019
\item \strong{Gebruiksrechten:} Onbeperkt
\item \strong{Soortgroep:} Macrofyten
}
}

\section{Bewerkingen}{


De originele kolomnaam abundantieklasse bevat een spelfout. Deze is gecorrigeerd.
De kolomnamen zijn omgezet naar lowercase.
}

\examples{

head(krw_mafy_score)
}
\seealso{
Zie ook \link{krw_mafy_categorie} en \link{krw_mafy_constanten}.
}
\keyword{datasets}
