% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_twn_opzoeklijsten_memoise.R
\name{opzoeklijsten}
\alias{opzoeklijsten}
\alias{fun_twn_voorkeur}
\alias{fun_twn_parent}
\alias{fun_twn_status}
\alias{fun_twn_status_tekst}
\alias{fun_twn_localname}
\alias{fun_twn_taxonlevel}
\title{Functies die opzoeklijsten maken}
\usage{
fun_twn_voorkeur()

fun_twn_parent()

fun_twn_status()

fun_twn_status_tekst()

fun_twn_localname()

fun_twn_taxonlevel()
}
\description{
Deze functies maken gebruik van de TWN-lijst en memoise om opzoeklijsten te maken t.b.v de
twn_info-functies. Uitsluitend voor intern gebruik
}
