% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZZ_package_description.R
\docType{package}
\name{twn}
\alias{twn}
\alias{twn-package}
\title{twn : Een package voor de TWN-lijst}
\description{
Package voor de TWN-lijst en bijbehorende functies. Zie verder: \url{https://redtent.github.io/twn/}
}
