% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twn_children.R
\name{twn_children}
\alias{twn_children}
\title{Vind alle 'child-taxa'}
\usage{
twn_children(
  parent,
  taxonlevel = c("All", "Subspecies", "Species", "Genus", "Familia", "Ordo", "Classis",
    "Phylum"),
  only_preferred = TRUE
)
}
\arguments{
\item{parent}{charactervector met naam (of namen) van het parenttaxon}

\item{taxonlevel}{Het gewenste taxonlevel van de children of \code{"All"} voor alle taxonlevels.}

\item{only_preferred}{logical. Alleen voorkeursnamen? Default is \code{TRUE}.}
}
\value{
EEn vector met de namen van de child-taxa
}
\description{
Deze functie kan gebruikt worden om alle child-taxa van een bepaalde parent te vinden.
}
\examples{

twn_children("Bufonidae")
twn_children("Bufonidae", taxonlevel = "Species")
twn_children("Bufonidae", only_preferred = FALSE)

}
