% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twn_children.R
\name{twn_children_rec}
\alias{twn_children_rec}
\title{Recursieve functie voor het opzoeken children}
\usage{
twn_children_rec(parent, child_tabel)
}
\arguments{
\item{parent}{De parentname: charactervector}

\item{child_tabel}{named vector opzoektabel voor children}
}
\value{
Vector met children
}
\description{
Recursieve functie voor het opzoeken children
}
\examples{

\dontrun{

twn_children_rec("Bufonidae", 
setNames(names(opzoektabel_twn_parent), opzoektabel_twn_parent))

}
}
\keyword{internal}
