% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twn_info.R
\name{twn_info}
\alias{twn_info}
\alias{twn_voorkeurnaam}
\alias{twn_parent}
\alias{twn_status}
\alias{twn_localname}
\alias{twn_taxonlevel}
\title{Opzoeken van de TWN informatie}
\usage{
twn_voorkeurnaam(namen)

twn_parent(namen)

twn_status(namen)

twn_localname(namen)

twn_taxonlevel(namen)
}
\arguments{
\item{namen}{Een vector met taxonnamen.}
}
\value{
Een vector met TWN-informatie. Namen die niet in de TWN worden
gevonden krijgen \code{NA}.
}
\description{
De \code{twn_*} - functies zoeken informatie uit de TWN-lijst op.
}
\details{
Deze functies accepteren een vector met taxonnamen en retourneren de
betreffende informatie uit de TWN-lijst. Als de taxonnaam niet in de
TWN-lijst voorkomt retourneert de functie \code{NA}
\itemize{
\item \code{twn_voorkeurnaam}: Geeft de voorkeurnaam van het taxon. Als het al de
voorkeurnaam is wordt de originele naam teruggegeven.
\item \code{twn_parent}: Geeft de parent van het taxon.
\item \code{twn_status}: Geeft de status van het taxon. Zie ook \link{twn_statuscodes}.
\item \code{twn_localname}: Geeft de Nederlandse naam van het taxon.
\item \code{twn_taxonlevel}: Geeft het taxonlevel van het taxon.
}
}
\note{
De TWN-lijst is niet altijd volledig consistent. Diverse taxa staan
meerdere malen in de lijst. In dat geval wordt het eerste resultaat
geretourneerd.

Als een taxon niet de voorkeurnaam is dan ontbreekt er vaak informatie
zoals de parent van het taxon. De functie \code{twn_parent} zoekt in dat
geval de parent van de voorkeurnaam.
}
\examples{
taxa <- c("Bufo calamita", "Bufo")

twn_voorkeurnaam(taxa)
twn_parent(taxa)
twn_status(taxa)
twn_localname(taxa)
twn_taxonlevel(taxa)

}
\seealso{
Deze functies werken op basis van de tabel \link{twn_lijst}
}
