% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twn_lijst_doc.R
\docType{data}
\name{twn_literatuur}
\alias{twn_literatuur}
\title{TWN literatuurlijst}
\format{
Dataframe met 3 kolommen
\itemize{
\item \code{literature}       - Literatuurcode conform de \code{twn_lijst}
\item \code{short_reference}  - Korte literatuurreferentie
\item \code{full_reference}   - Volledige literatuurreferentie
}
}
\source{
\url{https://twnlist.aquadesk.nl/}
}
\usage{
twn_literatuur
}
\description{
In de TWN-lijst worden de referenties naar de literatuur gegeven in codes. In deze lijst zijn alle
TWN-literatuurreferenties opgenomen.
}
\note{
De centrale TWN-literatuurlijst wordt regelmatig van updates voorzien. Deze
TWN-literatuurlijst kan daarom enigszins verouderd zijn. De datum van de gebruikte
TWN-literatuurlijst is opgeslagen in het attribuut \code{datum_twn_literatuur} en kan
worden opgevraagd met de code \code{attr(twn_literatuur, "datum_twn_literatuur")}.
}
\examples{

head(twn_literatuur)

# de datum van de gebruikte TWN-literatuurlijst.
attr(twn_literatuur, "datum_twn_literatuur")


}
\keyword{datasets}
