% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twn_lijst_doc.R
\docType{data}
\name{twn_statuscodes}
\alias{twn_statuscodes}
\title{TWN statuscodes}
\format{
Dataframe met 2 kolommen
\itemize{
\item \code{status}  - TWN-statuscode
\item \code{omschrijving}- Omschrijving van de betekenis van de statuscode
}
}
\usage{
twn_statuscodes
}
\description{
De TWN-lijst heeft aan ieder taxon een statuscode toegekend. De betekenis van deze codes is opgenomen in
de tabel \code{twn_statuscodes}
}
\examples{

twn_statuscodes

}
\keyword{datasets}
