% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twn_lijst_doc.R
\docType{data}
\name{twn_taxonlevels}
\alias{twn_taxonlevels}
\title{TWN taxonlevels}
\format{
Geordende factor
}
\usage{
twn_taxonlevels
}
\description{
De TWN-lijst heeft aan ieder taxon een taxonlevel toegekend. \code{twn_taxonlevels} is een
geordende factor met alle beschikbare taxonlevels.
}
\examples{

twn_taxonlevels

}
\keyword{datasets}
