% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wew_mafa_doc.R
\docType{data}
\name{wew_mafa_toelichting}
\alias{wew_mafa_toelichting}
\title{Toelichting op de betekenis van de indicatoren van wew_mafa_data}
\format{Dataframe
\itemize{
\item \code{indicator_groep} Het type milieufactor
\item \code{indicator_waarde_code}  Code voor de milieufactor
\item \code{indicator_waarde} Waarde van de milieufactor
\item \code{omschrijving} Betekenis van de indicatorwaarde
}}
\usage{
wew_mafa_toelichting
}
\description{
Deze dataset bevat een toelichting op de betekenis van de indicatorwaarden in \link{wew_mafa_data}.
}
\examples{

head(wew_mafa_toelichting)
}
\seealso{
\link{wew_mafa_data} en \link{wew_mafa_zeldzaamheid}
}
\keyword{datasets}
