% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wew_mafa_doc.R
\docType{data}
\name{wew_mafa_zeldzaamheid}
\alias{wew_mafa_zeldzaamheid}
\title{Zeldzaamheid van macrofauna}
\format{Dataframe
\itemize{
\item \code{taxonnaam_orig} Taxonnaam uit originele dataset
\item \code{taxonnaam_voorkeur} Voorkeursnaam taxon. Als deze ontbreekt betekent dat dat de \code{taxonnaam_orig}
niet in de TWN-lijst voorkomt.
\item \code{zeldzaamheid} Code voor de zeldzaamheid
\item \code{omschrijving} Beschrijving van de zeldzaamheid
}}
\source{
\url{http://www.wew.nu/publicaties/wew_themanummer23_soortenlijst.xls}

\url{http://www.wew.nu/publicaties/wew_themanummer23.pdf}
}
\usage{
wew_mafa_zeldzaamheid
}
\description{
Deze dataset bevat de zeldzaamheid van Nederlandse macrofaunasoorten en is opgesteld door
de subgroep Autecologie van de \href{www.wew.nu}{Werkgroep Ecologisch Waterbeheer}.
}
\note{
De zeldzaamheidscode is in sommige gevallen de code \code{u}. Deze code is in de documentatie niet gedefinieerd.
Aangenomen wordt dat deze code betekent dat de betreffende soort in Nederland is uitgestorven.
}
\section{Metadata dataset}{

\itemize{
\item \strong{Naam:} Milieu- en habitatpreferenties van Nederlandse zoetwatermacrofauna
\item \strong{Bronhouder:} Werkgroep Ecologisch Waterbeheer
\item \strong{Auteurs:}
\itemize{
\item \strong{Auteurs:} W.C.E.P. Verberk, P.F.M. Verdonschot, T. van Haaren, B. van Maanen.
\item \strong{Redactie:} Rebi C. Nijboer.
\item \emph{Subgroep Autoecologie, WEW-23.}
}
\item \strong{Datum dataset:} Mei 2012.
\item \strong{Datum opname dataset:} 10 April 2019
\item \strong{Gebruiksrechten:} Geen beperkingen
\item \strong{Soortgroep:} Macrofauna
}
}

\examples{

head(wew_mafa_zeldzaamheid)
}
\seealso{
\link{wew_mafa_data} en \link{wew_mafa_toelichting}
}
\keyword{datasets}
